refactor(iroh): simplify and cleanup the sync algorithm
switches to also use tokio::codec under the hood now
