+++ iroh/Cargo.toml
@@ -40,7 +40,7 @@ serde = { version = "1", features = ["derive"] }
 thiserror = "1"
 tokio = { version = "1", features = ["io-util", "rt"] }
 tokio-stream = "0.1"
-tokio-util = { version = "0.7", features = ["io-util", "io"] }
+tokio-util = { version = "0.7", features = ["codec", "io-util", "io"] }
 tracing = "0.1"
 walkdir = "2"
